introducing AbstractAssignToJudgeCorresponder.java
