+++ src/main/java/uk/gov/hmcts/reform/finrem/caseorchestration/service/correspondence/assigntojudge/AbstractAssignToJudgeCorresponder.java
@@ -1,0 +1,65 @@
+package uk.gov.hmcts.reform.finrem.caseorchestration.service.correspondence.assigntojudge;
+
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Component;
+import uk.gov.hmcts.reform.finrem.caseorchestration.helper.DocumentHelper;
+import uk.gov.hmcts.reform.finrem.caseorchestration.model.ccd.CaseDocument;
+import uk.gov.hmcts.reform.finrem.caseorchestration.model.ccd.FinremCaseDetails;
+import uk.gov.hmcts.reform.finrem.caseorchestration.model.ccd.wrapper.intevener.IntervenerWrapper;
+import uk.gov.hmcts.reform.finrem.caseorchestration.service.AssignedToJudgeDocumentService;
+import uk.gov.hmcts.reform.finrem.caseorchestration.service.BulkPrintService;
+import uk.gov.hmcts.reform.finrem.caseorchestration.service.NotificationService;
+import uk.gov.hmcts.reform.finrem.caseorchestration.service.correspondence.FinremSingleLetterOrEmailAllPartiesCorresponder;
+
+@Component
+@Slf4j
+public abstract class AbstractAssignToJudgeCorresponder extends FinremSingleLetterOrEmailAllPartiesCorresponder {
+    private final AssignedToJudgeDocumentService assignedToJudgeDocumentService;
+
+    @Autowired
+    public AbstractAssignToJudgeCorresponder(NotificationService notificationService,
+                                             BulkPrintService bulkPrintService,
+                                             AssignedToJudgeDocumentService assignedToJudgeDocumentService) {
+        super(notificationService, bulkPrintService);
+        this.assignedToJudgeDocumentService = assignedToJudgeDocumentService;
+    }
+
+    @Override
+    protected boolean shouldSendIntervenerLetter(IntervenerWrapper intervenerWrapper) {
+        return intervenerWrapper.getIntervenerName() != null && !intervenerWrapper.getIntervenerName().isEmpty();
+    }
+
+    @Override
+    protected boolean shouldSendRespondentSolicitorEmail(FinremCaseDetails caseDetails) {
+        return notificationService.isRespondentSolicitorEmailPopulated(caseDetails);
+    }
+
+    @Override
+    protected boolean shouldSendApplicantSolicitorEmail(FinremCaseDetails caseDetails) {
+        return notificationService.isApplicantSolicitorEmailPopulated(caseDetails);
+    }
+
+    @Override
+    public CaseDocument getDocumentToPrint(FinremCaseDetails caseDetails, String authorisationToken,
+                                           DocumentHelper.PaperNotificationRecipient recipient) {
+        return assignedToJudgeDocumentService.generateAssignedToJudgeNotificationLetter(
+            caseDetails, authorisationToken, recipient);
+    }
+
+    @Override
+    protected void emailApplicantSolicitor(FinremCaseDetails caseDetails) {
+        notificationService.sendAssignToJudgeConfirmationEmailToApplicantSolicitor(caseDetails);
+    }
+
+    @Override
+    protected void emailRespondentSolicitor(FinremCaseDetails caseDetails) {
+        notificationService.sendAssignToJudgeConfirmationEmailToRespondentSolicitor(caseDetails);
+    }
+
+    @Override
+    protected void emailIntervenerSolicitor(IntervenerWrapper intervenerWrapper, FinremCaseDetails caseDetails) {
+        notificationService.sendAssignToJudgeConfirmationEmailToIntervenerSolicitor(caseDetails,
+            notificationService.getCaseDataKeysForIntervenerSolicitor(intervenerWrapper));
+    }
+}
