+++ src/main/java/uk/gov/hmcts/reform/finrem/caseorchestration/service/correspondence/assigntojudge/FinremAssignToJudgeCorresponder.java
@@ -3,63 +3,19 @@ package uk.gov.hmcts.reform.finrem.caseorchestration.service.correspondence.assi
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
-import uk.gov.hmcts.reform.finrem.caseorchestration.helper.DocumentHelper;
-import uk.gov.hmcts.reform.finrem.caseorchestration.model.ccd.CaseDocument;
-import uk.gov.hmcts.reform.finrem.caseorchestration.model.ccd.FinremCaseDetails;
-import uk.gov.hmcts.reform.finrem.caseorchestration.model.ccd.wrapper.intevener.IntervenerWrapper;
 import uk.gov.hmcts.reform.finrem.caseorchestration.service.AssignedToJudgeDocumentService;
 import uk.gov.hmcts.reform.finrem.caseorchestration.service.BulkPrintService;
 import uk.gov.hmcts.reform.finrem.caseorchestration.service.NotificationService;
-import uk.gov.hmcts.reform.finrem.caseorchestration.service.correspondence.FinremSingleLetterOrEmailAllPartiesCorresponder;
 
 @Slf4j
 @Component
-public class FinremAssignToJudgeCorresponder extends FinremSingleLetterOrEmailAllPartiesCorresponder {
-    private final AssignedToJudgeDocumentService assignedToJudgeDocumentService;
+public class FinremAssignToJudgeCorresponder extends AbstractAssignToJudgeCorresponder {
 
     @Autowired
     public FinremAssignToJudgeCorresponder(NotificationService notificationService,
                                            BulkPrintService bulkPrintService,
                                            AssignedToJudgeDocumentService assignedToJudgeDocumentService) {
-        super(notificationService, bulkPrintService);
-        this.assignedToJudgeDocumentService = assignedToJudgeDocumentService;
+        super(notificationService, bulkPrintService, assignedToJudgeDocumentService);
     }
 
-    @Override
-    protected boolean shouldSendIntervenerLetter(IntervenerWrapper intervenerWrapper) {
-        return intervenerWrapper.getIntervenerName() != null && !intervenerWrapper.getIntervenerName().isEmpty();
-    }
-
-    @Override
-    protected boolean shouldSendRespondentSolicitorEmail(FinremCaseDetails caseDetails) {
-        return notificationService.isRespondentSolicitorEmailPopulated(caseDetails);
-    }
-
-    @Override
-    protected boolean shouldSendApplicantSolicitorEmail(FinremCaseDetails caseDetails) {
-        return notificationService.isApplicantSolicitorEmailPopulated(caseDetails);
-    }
-
-    @Override
-    public CaseDocument getDocumentToPrint(FinremCaseDetails caseDetails, String authorisationToken,
-                                           DocumentHelper.PaperNotificationRecipient recipient) {
-        return assignedToJudgeDocumentService.generateAssignedToJudgeNotificationLetter(
-            caseDetails, authorisationToken, recipient);
-    }
-
-    @Override
-    protected void emailApplicantSolicitor(FinremCaseDetails caseDetails) {
-        notificationService.sendAssignToJudgeConfirmationEmailToApplicantSolicitor(caseDetails);
-    }
-
-    @Override
-    protected void emailRespondentSolicitor(FinremCaseDetails caseDetails) {
-        notificationService.sendAssignToJudgeConfirmationEmailToRespondentSolicitor(caseDetails);
-    }
-
-    @Override
-    protected void emailIntervenerSolicitor(IntervenerWrapper intervenerWrapper, FinremCaseDetails caseDetails) {
-        notificationService.sendAssignToJudgeConfirmationEmailToIntervenerSolicitor(caseDetails,
-            notificationService.getCaseDataKeysForIntervenerSolicitor(intervenerWrapper));
-    }
 }
